changed rating from integer to float

+++ resources/migrations/2016_04_26_015845_create_user_ratings_table.php
@@ -21,7 +21,7 @@ class CreateUserRatingsTable extends Migration
             $table->boolean('like')->default(false);
             $table->boolean('dislike')->default(false);
             $table->boolean('favorite')->default(false);
-            $table->integer('rating')->nullable()->default(null);
+            $table->float('rating')->nullable()->default(null);
             $table->timestamps();
 
             $table->unique(array('namespace', 'entity_id', 'user_id'));
